remove airbnb eslint configs
After upgrading to the latest version of Yarn in multiple private
dependencies and removing the ../packages link infrastructure, I started
running into an error within the app that was complaining that
eslint-plugin-jsx-a11y was being imported multiple times.

Indeed, both eslint-config-airbnb and eslint-config-react-app both bring
in this plugin. It's likely neater to stick to one of these or converge
onto a single set of rules that we agree on. In either case, I need to
remove one of them to get the app to build now, so this patch removes
Airbnb's configs.